Add cron expr validation and reminder removal
Also all the id types become int64

+++ reminder/services/triggered_reminder.go
@@ -9,6 +9,6 @@ func GetTriggeredReminders(app *app.Application) []models.Reminder {
 	return app.RemindManager.GetReminds()
 }
 
-func CompleteReminds(app *app.Application, ids []int) {
+func CompleteReminds(app *app.Application, ids []int64) {
 	app.RemindManager.CompleteReminds(ids...)
 }
